disk_analyer: need to pass properly sized int

+++ src/disk_analyzer/mbr/mbr.c
@@ -84,7 +84,7 @@ int print_partition(struct partition_table_entry pte)
     fprintf_green(stdout, "First Sector LBA: 0x%.8"
                     PRIx32"\n",
                     pte.first_sector_lba);
-    pretty_print_bytes(SECTOR_SIZE*pte.sector_count, size_buf, 512);
+    pretty_print_bytes((uint64_t) SECTOR_SIZE*pte.sector_count, size_buf, 512);
     fprintf_green(stdout, "Number of Sectors: 0x%.8"
                     PRIx32" (%s)\n",
                     pte.sector_count,
